+++ src/inheritAttributes.js
@@ -1,71 +1,258 @@
 import { flatten } from './utils/list';
 import { shallowMerge, getAttributes } from './utils/object';
 import { parseDuration } from './utils/time';
-import resolveUrl from './resolveUrl';
+import { findChildren, getContent } from './utils/xml';
+import resolveUrl from './utils/resolveUrl';
 import errors from './errors';
+import merge from 'deepmerge';
 
-export const rep = mpdAttributes => (period, periodIndex) => {
-  const adaptationSets = Array.from(period.getElementsByTagName('AdaptationSet'));
+/**
+ * Builds a list of urls that is the product of the reference urls and BaseURL values
+ *
+ * @param {string[]} referenceUrls
+ *        List of reference urls to resolve to
+ * @param {Node[]} baseUrlElements
+ *        List of BaseURL nodes from the mpd
+ * @return {string[]}
+ *         List of resolved urls
+ */
+export const buildBaseUrls = (referenceUrls, baseUrlElements) => {
+  if (!baseUrlElements.length) {
+    return referenceUrls;
+  }
 
-  const representationsByAdaptationSet = adaptationSets.map(adaptationSet => {
-    const adaptationSetAttributes = getAttributes(adaptationSet);
+  return flatten(
+    referenceUrls.map(
+      reference => baseUrlElements.map(
+        baseUrlElement => resolveUrl(reference, getContent(baseUrlElement)))));
+};
 
-    const role = adaptationSet.getElementsByTagName('Role')[0];
-    const roleAttributes = { role: getAttributes(role) };
+/**
+ * Contains all Segment information for its containing AdaptationSet
+ *
+ * @typedef {Object} SegmentInformation
+ * @property {Object|undefined} template
+ *           Contains the attributes for the SegmentTemplate node
+ * @property {Object[]|undefined} timeline
+ *           Contains a list of atrributes for each S node within the SegmentTimeline node
+ * @property {Object|undefined} list
+ *           Contains the attributes for the SegmentList node
+ * @property {Object|undefined} base
+ *           Contains the attributes for the SegmentBase node
+ */
 
-    const attrs = shallowMerge({ periodIndex }, mpdAttributes, adaptationSetAttributes, roleAttributes);
+/**
+ * Returns all available Segment information contained within the AdaptationSet node
+ *
+ * @param {Node} adaptationSet
+ *        The AdaptationSet node to get Segment information from
+ * @return {SegmentInformation}
+ *         The Segment information contained within the provided AdaptationSet
+ */
+export const getSegmentInformation = (adaptationSet) => {
+  const segmentTemplate = findChildren(adaptationSet, 'SegmentTemplate')[0];
+  const segmentList = findChildren(adaptationSet, 'SegmentList')[0];
+  const segmentUrls = segmentList && findChildren(segmentList, 'SegmentURL')
+    .map(s => shallowMerge({ tag: 'SegmentURL' }, getAttributes(s)));
+  const segmentBase = findChildren(adaptationSet, 'SegmentBase')[0];
+  const segmentTimelineParentNode = segmentList || segmentTemplate;
+  const segmentTimeline = segmentTimelineParentNode &&
+    findChildren(segmentTimelineParentNode, 'SegmentTimeline')[0];
+  const segmentInitializationParentNode = segmentList || segmentBase || segmentTemplate;
+  const segmentInitialization = segmentInitializationParentNode &&
+    findChildren(segmentInitializationParentNode, 'Initialization')[0];
 
-    const segmentTemplate = adaptationSet.getElementsByTagName('SegmentTemplate')[0];
-    const segmentList = adaptationSet.getElementsByTagName('SegmentList')[0];
-    const segmentBase = adaptationSet.getElementsByTagName('SegmentBase')[0];
+  // SegmentTemplate is handled slightly differently, since it can have both
+  // @initialization and an <Initialization> node.  @initialization can be templated,
+  // while the node can have a url and range specified.  If the <SegmentTemplate> has
+  // both @initialization and an <Initialization> subelement we opt to override with
+  // the node, as this interaction is not defined in the spec.
+  const template = segmentTemplate && getAttributes(segmentTemplate);
 
-    const segmentType = {
-      segmentTemplate: segmentTemplate && getAttributes(segmentTemplate),
-      segmentList: segmentList && getAttributes(segmentList),
-      segmentBase: segmentBase && getAttributes(segmentBase)
-    };
+  if (template && segmentInitialization) {
+    template.initialization =
+      (segmentInitialization && getAttributes(segmentInitialization));
+  } else if (template && template.initialization) {
+    // If it is @initialization we convert it to an object since this is the format that
+    // later functions will rely on for the initialization segment.  This is only valid
+    // for <SegmentTemplate>
+    template.initialization = { sourceURL: template.initialization };
+  }
 
-    const representations = Array.from(adaptationSet.getElementsByTagName('Representation'));
+  const segmentInfo = {
+    template,
+    timeline: segmentTimeline &&
+      findChildren(segmentTimeline, 'S').map(s => getAttributes(s)),
+    list: segmentList && shallowMerge(
+      getAttributes(segmentList),
+      {
+        segmentUrls,
+        initialization: getAttributes(segmentInitialization)
+      }),
+    base: segmentBase && shallowMerge(
+      getAttributes(segmentBase), {
+        initialization: getAttributes(segmentInitialization)
+      })
+  };
 
-    const inherit = representation => {
-      // vtt tracks may use single file in BaseURL
-      let baseUrl = representation.getElementsByTagName('BaseURL')[0];
-
-      baseUrl = baseUrl && baseUrl.innerHTML || '';
-      const attributes = shallowMerge(attrs, getAttributes(representation), { url: baseUrl });
-
-      return { attributes, segmentType };
-    };
-
-    return representations.map(inherit);
+  Object.keys(segmentInfo).forEach(key => {
+    if (!segmentInfo[key]) {
+      delete segmentInfo[key];
+    }
   });
 
-  return flatten(representationsByAdaptationSet);
+  return segmentInfo;
 };
 
-export const representationsByPeriod = (periods, mpdAttributes) => {
-  return periods.map(rep(mpdAttributes));
+/**
+ * Contains Segment information and attributes needed to construct a Playlist object
+ * from a Representation
+ *
+ * @typedef {Object} RepresentationInformation
+ * @property {SegmentInformation} segmentInfo
+ *           Segment information for this Representation
+ * @property {Object} attributes
+ *           Inherited attributes for this Representation
+ */
+
+/**
+ * Maps a Representation node to an object containing Segment information and attributes
+ *
+ * @name inheritBaseUrlsCallback
+ * @function
+ * @param {Node} representation
+ *        Representation node from the mpd
+ * @return {RepresentationInformation}
+ *         Representation information needed to construct a Playlist object
+ */
+
+/**
+ * Returns a callback for Array.prototype.map for mapping Representation nodes to
+ * Segment information and attributes using inherited BaseURL nodes.
+ *
+ * @param {Object} adaptationSetAttributes
+ *        Contains attributes inherited by the AdaptationSet
+ * @param {string[]} adaptationSetBaseUrls
+ *        Contains list of resolved base urls inherited by the AdaptationSet
+ * @param {SegmentInformation} segmentInfo
+ *        Contains Segment information for the AdaptationSet
+ * @return {inheritBaseUrlsCallback}
+ *         Callback map function
+ */
+export const inheritBaseUrls =
+(adaptationSetAttributes, adaptationSetBaseUrls, adaptationSetSegmentInfo) => (representation) => {
+  const repBaseUrlElements = findChildren(representation, 'BaseURL');
+  const repBaseUrls = buildBaseUrls(adaptationSetBaseUrls, repBaseUrlElements);
+  const attributes = shallowMerge(adaptationSetAttributes, getAttributes(representation));
+  const representationSegmentInfo = getSegmentInformation(representation);
+
+  return repBaseUrls.map(baseUrl => {
+    return {
+      segmentInfo: merge(adaptationSetSegmentInfo, representationSegmentInfo),
+      attributes: shallowMerge(attributes, { baseUrl })
+    };
+  });
 };
 
+/**
+ * Maps an AdaptationSet node to a list of Representation information objects
+ *
+ * @name toRepresentationsCallback
+ * @function
+ * @param {Node} adaptationSet
+ *        AdaptationSet node from the mpd
+ * @return {RepresentationInformation[]}
+ *         List of objects containing Representaion information
+ */
+
+/**
+ * Returns a callback for Array.prototype.map for mapping AdaptationSet nodes to a list of
+ * Representation information objects
+ *
+ * @param {Object} periodAttributes
+ *        Contains attributes inherited by the Period
+ * @param {string[]} periodBaseUrls
+ *        Contains list of resolved base urls inherited by the Period
+ * @return {toRepresentationsCallback}
+ *         Callback map function
+ */
+export const toRepresentations =
+(periodAttributes, periodBaseUrls, periodSegmentInfo) => (adaptationSet) => {
+  const adaptationSetAttributes = getAttributes(adaptationSet);
+  const adaptationSetBaseUrls = buildBaseUrls(periodBaseUrls,
+                                              findChildren(adaptationSet, 'BaseURL'));
+  const role = findChildren(adaptationSet, 'Role')[0];
+  const roleAttributes = { role: getAttributes(role) };
+  const attrs = shallowMerge(periodAttributes,
+                             adaptationSetAttributes,
+                             roleAttributes);
+  const segmentInfo = getSegmentInformation(adaptationSet);
+  const representations = findChildren(adaptationSet, 'Representation');
+  const adaptationSetSegmentInfo = merge(periodSegmentInfo, segmentInfo);
+
+  return flatten(
+    representations.map(inheritBaseUrls(attrs, adaptationSetBaseUrls, adaptationSetSegmentInfo)));
+};
+
+/**
+ * Maps an Period node to a list of Representation inforamtion objects for all
+ * AdaptationSet nodes contained within the Period
+ *
+ * @name toAdaptationSetsCallback
+ * @function
+ * @param {Node} period
+ *        Period node from the mpd
+ * @param {number} periodIndex
+ *        Index of the Period within the mpd
+ * @return {RepresentationInformation[]}
+ *         List of objects containing Representaion information
+ */
+
+/**
+ * Returns a callback for Array.prototype.map for mapping Period nodes to a list of
+ * Representation information objects
+ *
+ * @param {Object} mpdAttributes
+ *        Contains attributes inherited by the mpd
+ * @param {string[]} mpdBaseUrls
+ *        Contains list of resolved base urls inherited by the mpd
+ * @return {toAdaptationSetsCallback}
+ *         Callback map function
+ */
+export const toAdaptationSets = (mpdAttributes, mpdBaseUrls) => (period, periodIndex) => {
+  const periodBaseUrls = buildBaseUrls(mpdBaseUrls, findChildren(period, 'BaseURL'));
+  const periodAtt = getAttributes(period);
+  const periodAttributes = shallowMerge(mpdAttributes, periodAtt, { periodIndex });
+  const adaptationSets = findChildren(period, 'AdaptationSet');
+  const periodSegmentInfo = getSegmentInformation(period);
+
+  return flatten(adaptationSets.map(toRepresentations(periodAttributes, periodBaseUrls, periodSegmentInfo)));
+};
+
+/**
+ * Traverses the mpd xml tree to generate a list of Representation information objects
+ * that have inherited attributes from parent nodes
+ *
+ * @param {Node} mpd
+ *        The root node of the mpd
+ * @param {string} manifestUri
+ *        The uri of the source mpd
+ * @return {RepresentationInformation[]}
+ *         List of objects containing Representation information
+ */
 export const inheritAttributes = (mpd, manifestUri = '') => {
-  const periods = Array.from(mpd.getElementsByTagName('Period'));
+  const periods = findChildren(mpd, 'Period');
 
-  if (!periods.length ||
-      periods.length &&
-      periods.length !== 1) {
+  if (periods.length !== 1) {
     // TODO add support for multiperiod
     throw new Error(errors.INVALID_NUMBER_OF_PERIOD);
   }
 
   const mpdAttributes = getAttributes(mpd);
-  const BaseUrl = mpd.getElementsByTagName('BaseURL');
+  const mpdBaseUrls = buildBaseUrls([ manifestUri ], findChildren(mpd, 'BaseURL'));
 
-  const baseUrl = BaseUrl && BaseUrl.length ? BaseUrl[0].innerHTML : '';
-
-  mpdAttributes.baseUrl = resolveUrl(manifestUri, baseUrl);
   mpdAttributes.sourceDuration = mpdAttributes.mediaPresentationDuration ?
     parseDuration(mpdAttributes.mediaPresentationDuration) : 0;
 
-  return flatten(representationsByPeriod(periods, mpdAttributes));
+  return flatten(periods.map(toAdaptationSets(mpdAttributes, mpdBaseUrls)));
 };
-