+++ test/toM3u8.test.js
@@ -43,9 +43,8 @@ QUnit.test('playlists', function(assert) {
       bandwidth: '20000',
       periodIndex: 1,
       mimeType: 'text/vtt',
-      url: 'https://www.example.com/vtt'
-    },
-    segments: []
+      baseUrl: 'https://www.example.com/vtt'
+    }
   }, {
     attributes: {
       sourceDuration: 100,
@@ -53,9 +52,8 @@ QUnit.test('playlists', function(assert) {
       bandwidth: '10000',
       periodIndex: 1,
       mimeType: 'text/vtt',
-      url: 'https://www.example.com/vtt'
-    },
-    segments: []
+      baseUrl: 'https://www.example.com/vtt'
+    }
   }];
 
   const expected = {
@@ -102,7 +100,12 @@ QUnit.test('playlists', function(assert) {
               },
               endList: true,
               resolvedUri: 'https://www.example.com/vtt',
-              segments: [],
+              segments: [{
+                duration: 100,
+                resolvedUri: 'https://www.example.com/vtt',
+                timeline: 1,
+                uri: 'https://www.example.com/vtt'
+              }],
               timeline: 1,
               uri: ''
             }],
@@ -115,6 +118,7 @@ QUnit.test('playlists', function(assert) {
     playlists: [{
       attributes: {
         AUDIO: 'audio',
+        SUBTITLES: 'subs',
         BANDWIDTH: 10000,
         CODECS: 'foo;bar',
         NAME: '1',
@@ -127,6 +131,285 @@ QUnit.test('playlists', function(assert) {
       endList: true,
       resolvedUri: '',
       segments: [],
+      timeline: 1,
+      uri: ''
+    }],
+    segments: [],
+    uri: ''
+  };
+
+  assert.deepEqual(toM3u8(input), expected);
+});
+
+QUnit.test('playlists with segments', function(assert) {
+  const input = [{
+    attributes: {
+      id: '1',
+      codecs: 'foo;bar',
+      sourceDuration: 100,
+      bandwidth: '20000',
+      periodIndex: 1,
+      mimeType: 'audio/mp4'
+    },
+    segments: [{
+      uri: '',
+      timeline: 1,
+      duration: 2,
+      resolvedUri: '',
+      map: {
+        uri: '',
+        resolvedUri: ''
+      }
+    }, {
+      uri: '',
+      timeline: 1,
+      duration: 2,
+      resolvedUri: '',
+      map: {
+        uri: '',
+        resolvedUri: ''
+      }
+    }]
+  }, {
+    attributes: {
+      id: '2',
+      codecs: 'foo;bar',
+      sourceDuration: 100,
+      bandwidth: '10000',
+      periodIndex: 1,
+      mimeType: 'audio/mp4'
+    },
+    segments: [{
+      uri: '',
+      timeline: 1,
+      duration: 2,
+      resolvedUri: '',
+      map: {
+        uri: '',
+        resolvedUri: ''
+      }
+    }, {
+      uri: '',
+      timeline: 1,
+      duration: 2,
+      resolvedUri: '',
+      map: {
+        uri: '',
+        resolvedUri: ''
+      }
+    }]
+  }, {
+    attributes: {
+      sourceDuration: 100,
+      id: '1',
+      width: '800',
+      height: '600',
+      codecs: 'foo;bar',
+      bandwidth: '10000',
+      periodIndex: 1,
+      mimeType: 'video/mp4'
+    },
+    segments: [{
+      uri: '',
+      timeline: 1,
+      duration: 2,
+      resolvedUri: '',
+      map: {
+        uri: '',
+        resolvedUri: ''
+      }
+    }, {
+      uri: '',
+      timeline: 1,
+      duration: 2,
+      resolvedUri: '',
+      map: {
+        uri: '',
+        resolvedUri: ''
+      }
+    }]
+  }, {
+    attributes: {
+      sourceDuration: 100,
+      id: '1',
+      bandwidth: '20000',
+      periodIndex: 1,
+      mimeType: 'text/vtt',
+      baseUrl: 'https://www.example.com/vtt'
+    },
+    segments: [{
+      uri: '',
+      timeline: 1,
+      duration: 2,
+      resolvedUri: '',
+      map: {
+        uri: '',
+        resolvedUri: ''
+      }
+    }, {
+      uri: '',
+      timeline: 1,
+      duration: 2,
+      resolvedUri: '',
+      map: {
+        uri: '',
+        resolvedUri: ''
+      }
+    }]
+  }, {
+    attributes: {
+      sourceDuration: 100,
+      id: '1',
+      bandwidth: '10000',
+      periodIndex: 1,
+      mimeType: 'text/vtt',
+      baseUrl: 'https://www.example.com/vtt'
+    },
+    segments: [{
+      uri: '',
+      timeline: 1,
+      duration: 2,
+      resolvedUri: '',
+      map: {
+        uri: '',
+        resolvedUri: ''
+      }
+    }, {
+      uri: '',
+      timeline: 1,
+      duration: 2,
+      resolvedUri: '',
+      map: {
+        uri: '',
+        resolvedUri: ''
+      }
+    }]
+  }];
+
+  const expected = {
+    allowCache: true,
+    discontinuityStarts: [],
+    duration: 100,
+    endList: true,
+    mediaGroups: {
+      AUDIO: {
+        audio: {
+          main: {
+            autoselect: true,
+            default: true,
+            language: '',
+            playlists: [{
+              attributes: {
+                BANDWIDTH: 20000,
+                CODECS: 'foo;bar',
+                NAME: '1',
+                ['PROGRAM-ID']: 1
+              },
+              endList: true,
+              resolvedUri: '',
+              segments: [{
+                uri: '',
+                timeline: 1,
+                duration: 2,
+                resolvedUri: '',
+                map: {
+                  uri: '',
+                  resolvedUri: ''
+                }
+              }, {
+                uri: '',
+                timeline: 1,
+                duration: 2,
+                resolvedUri: '',
+                map: {
+                  uri: '',
+                  resolvedUri: ''
+                }
+              }],
+              timeline: 1,
+              uri: ''
+            }],
+            uri: ''
+          }
+        }
+      },
+      ['CLOSED-CAPTIONS']: {},
+      SUBTITLES: {
+        subs: {
+          text: {
+            autoselect: false,
+            default: false,
+            language: 'text',
+            playlists: [{
+              attributes: {
+                BANDWIDTH: 20000,
+                NAME: '1',
+                ['PROGRAM-ID']: 1
+              },
+              endList: true,
+              resolvedUri: 'https://www.example.com/vtt',
+              segments: [{
+                uri: '',
+                timeline: 1,
+                duration: 2,
+                resolvedUri: '',
+                map: {
+                  uri: '',
+                  resolvedUri: ''
+                }
+              }, {
+                uri: '',
+                timeline: 1,
+                duration: 2,
+                resolvedUri: '',
+                map: {
+                  uri: '',
+                  resolvedUri: ''
+                }
+              }],
+              timeline: 1,
+              uri: ''
+            }],
+            uri: ''
+          }
+        }
+      },
+      VIDEO: {}
+    },
+    playlists: [{
+      attributes: {
+        AUDIO: 'audio',
+        SUBTITLES: 'subs',
+        BANDWIDTH: 10000,
+        CODECS: 'foo;bar',
+        NAME: '1',
+        ['PROGRAM-ID']: 1,
+        RESOLUTION: {
+          height: 600,
+          width: 800
+        }
+      },
+      endList: true,
+      resolvedUri: '',
+      segments: [{
+        uri: '',
+        timeline: 1,
+        duration: 2,
+        resolvedUri: '',
+        map: {
+          uri: '',
+          resolvedUri: ''
+        }
+      }, {
+        uri: '',
+        timeline: 1,
+        duration: 2,
+        resolvedUri: '',
+        map: {
+          uri: '',
+          resolvedUri: ''
+        }
+      }],
       timeline: 1,
       uri: ''
     }],