+++ test/toPlaylists.test.js
@@ -1,8 +1,6 @@
 import {
-  toPlaylists,
-  segmentsFromTemplate
+  toPlaylists
 } from '../src/toPlaylists';
-import errors from '../src/errors';
 import QUnit from 'qunit';
 
 QUnit.module('toPlaylists');
@@ -11,62 +9,26 @@ QUnit.test('no representations', function(assert) {
   assert.deepEqual(toPlaylists([]), []);
 });
 
-QUnit.test('simple', function(assert) {
-  const attributes = {
-    startNumber: '0',
-    duration: '2000',
-    timescale: '1000',
-    id: 'id',
-    initialization: 'init.mp4',
-    sourceDuration: 6,
-    media: '$Number$.mp4',
-    periodIndex: '',
-    baseUrl: 'https://www.example.com/mpd/'
-  };
-
-  const segments = [{
-    duration: 2,
-    map: {
-      resolvedUri: 'https://www.example.com/mpd/init.mp4',
-      uri: 'init.mp4'
-    },
-    resolvedUri: 'https://www.example.com/mpd/0.mp4',
-    timeline: '',
-    uri: '0.mp4'
-  }, {
-    duration: 2,
-    map: {
-      resolvedUri: 'https://www.example.com/mpd/init.mp4',
-      uri: 'init.mp4'
-    },
-    resolvedUri: 'https://www.example.com/mpd/1.mp4',
-    timeline: '',
-    uri: '1.mp4'
-  }, {
-    duration: 2,
-    map: {
-      resolvedUri: 'https://www.example.com/mpd/init.mp4',
-      uri: 'init.mp4'
-    },
-    resolvedUri: 'https://www.example.com/mpd/2.mp4',
-    timeline: '',
-    uri: '2.mp4'
-  }];
-
-  assert.deepEqual(segmentsFromTemplate(attributes), segments);
-});
-
 QUnit.test('pretty simple', function(assert) {
   const representations = [{
-    attributes: {},
-    segmentType: {
-      segmentTemplate: true
+    attributes: { baseUrl: 'http://example.com/' },
+    segmentInfo: {
+      template: true
     }
   }];
 
   const playlists = [{
-    attributes: {},
-    segments: []
+    attributes: { baseUrl: 'http://example.com/' },
+    segments: [{
+      uri: '',
+      timeline: undefined,
+      duration: undefined,
+      resolvedUri: 'http://example.com/',
+      map: {
+        uri: '',
+        resolvedUri: 'http://example.com/'
+      }
+    }]
   }];
 
   assert.deepEqual(toPlaylists(representations), playlists);
@@ -74,24 +36,71 @@ QUnit.test('pretty simple', function(assert) {
 
 QUnit.test('segment base', function(assert) {
   const representations = [{
-    attributes: {},
-    segmentType: {
-      segmentBase: true
+    attributes: { baseUrl: 'http://example.com/' },
+    segmentInfo: {
+      base: true
     }
   }];
 
-  assert.throws(() => toPlaylists(representations),
-    new RegExp(errors.UNSUPPORTED_SEGMENTATION_TYPE));
+  const playlists = [{
+    attributes: { baseUrl: 'http://example.com/' },
+    segments: [{
+      map: {
+        resolvedUri: 'http://example.com/',
+        uri: ''
+      },
+      resolvedUri: 'http://example.com/',
+      uri: 'http://example.com/'
+    }]
+  }];
+
+  assert.deepEqual(toPlaylists(representations), playlists);
 });
 
 QUnit.test('segment list', function(assert) {
   const representations = [{
-    attributes: {},
-    segmentType: {
-      segmentList: true
+    attributes: {
+      baseUrl: 'http://example.com/',
+      duration: 10,
+      sourceDuration: 11
+    },
+    segmentInfo: {
+      list: {
+        segmentUrls: [{
+          media: '1.fmp4'
+        }, {
+          media: '2.fmp4'
+        }]
+      }
     }
   }];
 
-  assert.throws(() => toPlaylists(representations),
-    new RegExp(errors.UNSUPPORTED_SEGMENTATION_TYPE));
+  const playlists = [{
+    attributes: {
+      baseUrl: 'http://example.com/',
+      duration: 10,
+      sourceDuration: 11
+    },
+    segments: [{
+      duration: 10,
+      map: {
+        resolvedUri: 'http://example.com/',
+        uri: ''
+      },
+      resolvedUri: 'http://example.com/1.fmp4',
+      timeline: 0,
+      uri: '1.fmp4'
+    }, {
+      duration: 1,
+      map: {
+        resolvedUri: 'http://example.com/',
+        uri: ''
+      },
+      resolvedUri: 'http://example.com/2.fmp4',
+      timeline: 0,
+      uri: '2.fmp4'
+    }]
+  }];
+
+  assert.deepEqual(toPlaylists(representations), playlists);
 });